lexer

+++ utils/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: cbourajl <cbourajl@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 21:17:46 by abouazi           #+#    #+#             */
-/*   Updated: 2022/09/04 16:53:49 by cbourajl         ###   ########.fr       */
+/*   Updated: 2022/09/14 16:10:30 by cbourajl         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	c;
+	int	c;
 
 	c = 0;
 	while (s[c] != '\0')
